+++ apps/pet-shop/src/app/integrator/ui/ps-products/ps-products.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
-import { PSProductsComponent } from './products.component'
+import { PSProductsComponent } from './ps-products.component'
 
 // TODO: Explore the tests, refactor, {migrate to `vitest`?}.
