+++ apps/pet-shop/src/app/integrator/ui/ps-products/ps-products.component.ts
@@ -18,8 +18,8 @@ import {
 } from '@angular/material/table'
 
 // # Internal modules
-import { PSEmptinessComponent } from '../../integrator/ui-kit/ps-emptiness/ps-emptiness.component'
-import { ProductTableViewModel } from './shared/product-table-view.model'
+import type { ProductTableViewModel } from '../../../shop/products/shared/product-table-view.model'
+import { PSEmptinessComponent } from '../../ui-kit/ps-emptiness/ps-emptiness.component'
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
